Fixed animations and colors in basalt
The animation for every hour works more smoothly now and shows the text as it should be if it was covered by the disc without the need of being the same color.

More options to be configured on basalt, giving much more freedom to decide over a style.

Fixed a bug where at 12 the animation wouldn't work.