Added date and action on tap
Added a date with configuration options (format, position)

Added marks for minutes to see with more precision the time.

On tap it will show all data that is hidden for 5 seconds (configuration coming in the future). Between the data shown there is the battery level, bluetooth connection status, minute marks, seconds mark and date.